dom/base/test/test_bug793311.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=793311
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 793311</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug {793311} **/
  SimpleTest.waitForExplicitFinish();

  try {
    SpecialPowers.DOMWindowUtils.wrapDOMFile(null);
    ok(false, "wrapDOMFile(null) throws an exception");
  } catch(e) {
    ok(true, "wrapDOMFile(null) throws an exception");
  }
  SimpleTest.finish();

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=793311">Mozilla Bug 793311</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
